trim samples
