+++ examples/api/lib/cupertino/sheet/cupertino_sheet.2.dart
@@ -4,26 +4,104 @@
 
 import 'package:flutter/cupertino.dart';
 
-/// Flutter code sample for [showCupertinoSheet].
+/// Flutter code sample for [CupertinoSheetRoute] with restorable state and nested navigation.
 
-void main() {
-  runApp(const CupertinoSheetApp());
-}
+void main() => runApp(const RestorableSheetExampleApp());
 
-class CupertinoSheetApp extends StatelessWidget {
-  const CupertinoSheetApp({super.key});
+class RestorableSheetExampleApp extends StatelessWidget {
+  const RestorableSheetExampleApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const CupertinoApp(
-      title: 'Cupertino Sheet',
-      home: HomePage(),
+      restorationScopeId: 'sheet-app',
+      title: 'Restorable Sheet',
+      home: RestorableSheet(restorationId: 'sheet'),
     );
   }
 }
 
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
+class RestorableSheet extends StatefulWidget {
+  const RestorableSheet({super.key, this.restorationId});
+
+  final String? restorationId;
+
+  @override
+  State<RestorableSheet> createState() => _RestorableSheetState();
+}
+
+class _RestorableSheetState extends State<RestorableSheet>
+    with RestorationMixin {
+
+  final RestorableInt _counter = RestorableInt(0);
+  late RestorableRouteFuture<int?> _restorableSheetRouteFuture;
+
+  @override
+  void initState() {
+    super.initState();
+    _restorableSheetRouteFuture = RestorableRouteFuture<int?>(
+      onComplete: _changeCounter,
+      onPresent: (NavigatorState navigator, Object? arguments) {
+        return navigator.restorablePush(
+          _counterSheetBuilder,
+          arguments: _counter.value
+        );
+      },
+    );
+  }
+
+  @override
+  String? get restorationId => widget.restorationId;
+
+  @override
+  void restoreState(RestorationBucket? oldBucket, bool initialRestore) {
+    registerForRestoration(_counter, 'count');
+    registerForRestoration(_restorableSheetRouteFuture, 'sheet_route_future');
+  }
+
+  @override
+  void dispose() {
+    _counter.dispose();
+    super.dispose();
+  }
+
+  @pragma('vm:entry-point')
+  static Route<void> _counterSheetBuilder(BuildContext context, Object? arguments) {
+    return CupertinoSheetRoute<int?>(
+      builder: (BuildContext context)  {
+        return Navigator(
+          restorationScopeId: 'nested-nav',
+          onGenerateRoute: (RouteSettings settings) {
+            return CupertinoPageRoute<void>(
+              settings: settings,
+              builder: (BuildContext context) {
+                return PopScope(
+                  canPop: settings.name != '/',
+                  onPopInvokedWithResult: (bool didPop, Object? result) {
+                    if (didPop) {
+                      return;
+                    }
+                    Navigator.of(context).pop();
+                  },
+                  child: CounterSheetScaffold(
+                    counter: arguments! as int,
+                  )
+                );
+              }
+            );
+          },
+        );
+      },
+    );
+  }
+
+  void _changeCounter(int? newCounter) {
+    if (newCounter != null) {
+      setState(() {
+        _counter.value = newCounter;
+      });
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -36,16 +114,18 @@ class HomePage extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            CupertinoButton.filled(
-              onPressed: () {
-                showCupertinoSheet<void>(
-                  context: context,
-                  useNestedNavigation: true,
-                  pageBuilder: (BuildContext context) => const SheetScaffold(),
-                );
-              },
-              child: const Text('Open Bottom Sheet'),
+            const Text(
+              'Counter current value:',
             ),
+            Text(
+              '${_counter.value}',
+            ),
+            CupertinoButton(
+              child: const Text('Open Sheet'),
+              onPressed: () {
+                _restorableSheetRouteFuture.present();
+              }
+            )
           ],
         ),
       ),
@@ -53,76 +133,165 @@ class HomePage extends StatelessWidget {
   }
 }
 
-class SheetScaffold extends StatelessWidget {
-  const SheetScaffold({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const CupertinoPageScaffold(
-      child: SheetBody(title: 'CupertinoSheetRoute')
-    );
-  }
-}
-
-class SheetBody extends StatelessWidget {
-  const SheetBody({
+class CounterSheetScaffold extends StatefulWidget  {
+  const CounterSheetScaffold({
     super.key,
-    required this.title,
+    required this.counter,
   });
 
-  final String title;
+  final int counter;
+
+  @override
+  State<CounterSheetScaffold> createState() => _CounterSheetScaffoldState();
+}
+
+class _CounterSheetScaffoldState extends State<CounterSheetScaffold> with RestorationMixin {
+  // late final RestorableInt _counter = RestorableInt(widget.counter);
+  late RestorableInt _counter;
+  late RestorableRouteFuture<int?> _multiplicationRouteFuture;
+
+  @override
+  void initState() {
+    super.initState();
+    _counter = RestorableInt(widget.counter);
+    _multiplicationRouteFuture = RestorableRouteFuture<int?>(
+      onComplete: _changeCounter,
+      onPresent: (NavigatorState navigator, Object? arguments) {
+        return navigator.restorablePush(
+          _multiplicationRouteBuilder,
+          arguments: _counter.value
+        );
+      },
+    );
+  }
+
+
+  @pragma('vm:entry-point')
+  static Route<void> _multiplicationRouteBuilder(BuildContext context, Object? arguments) {
+    return CupertinoPageRoute<int?>(
+      settings: const RouteSettings(name: '/multiplication'),
+      builder: (BuildContext context) {
+        return MultiplicationPage(
+            counter: arguments! as int,
+        );
+      }
+    );
+  }
+
+  void _changeCounter(int? newCounter) {
+    if (newCounter != null) {
+      setState(() {
+        _counter.value = newCounter;
+      });
+    }
+  }
+
+  @override
+  String? get restorationId => 'sheet_scaffold';
+
+  @override
+  void restoreState(RestorationBucket? oldBucket, bool initialRestore) {
+    registerForRestoration(_counter, 'sheet_counter');
+    registerForRestoration(_multiplicationRouteFuture, 'multiplication_route');
+    if (!_counter.enabled) {
+      _counter = RestorableInt(widget.counter);
+    }
+  }
+
+  @override
+  void dispose() {
+    _counter.dispose();
+    _multiplicationRouteFuture.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Text(title),
-          CupertinoButton.filled(
-            onPressed: () {
-              Navigator.of(context).pop();
-            },
-            child: const Text('Go Back'),
-          ),
-          CupertinoButton.filled(
-            onPressed: () {
-              CupertinoSheetRoute.popSheet(context);
-            },
-            child: const Text('Pop Whole Sheet'),
-          ),
-          CupertinoButton.filled(
-            onPressed: () {
-              Navigator.of(context).push(
-                CupertinoPageRoute<void>(builder: (BuildContext context) => const SheetNextPage())
-              );
-            },
-            child: const Text('Push Nested Page'),
-          ),
-          CupertinoButton.filled(
-            onPressed: () {
-              showCupertinoSheet<void>(
-                context: context,
-                useNestedNavigation: true,
-                pageBuilder: (BuildContext context) => const SheetScaffold(),
-              );
-            },
-            child: const Text('Push Another Sheet'),
-          ),
-        ],
-      ),
+    return CupertinoPageScaffold
+    (
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text('Current Count'),
+            Text(_counter.value.toString()),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                CupertinoButton(
+                  onPressed: () {setState(() => _counter.value = _counter.value - 1);},
+                  child: const Text('Decrease'),
+                ),
+                CupertinoButton(
+                  onPressed: () {setState(() => _counter.value = _counter.value + 1);},
+                  child: const Text('Increase'),
+                ),
+              ],
+            ),
+            CupertinoButton(
+              onPressed: () => _multiplicationRouteFuture.present(),
+              child: const Text('Go to Multiplication Page'),
+            ),
+            CupertinoButton(
+              onPressed: () => Navigator.of(context, rootNavigator: true).pop(_counter.value),
+              child: const Text('Pop Sheet'),
+            ),
+          ]
+        ),
+      )
     );
   }
 }
 
-class SheetNextPage extends StatelessWidget {
-  const SheetNextPage({super.key});
+class MultiplicationPage extends StatefulWidget  {
+  const MultiplicationPage ({
+    super.key,
+    required this.counter,
+  });
+
+  final int counter;
+
+  @override
+  State<MultiplicationPage > createState() => _MultiplicationPageState();
+}
+
+class _MultiplicationPageState extends State<MultiplicationPage> with RestorationMixin {
+  late final RestorableInt _counter = RestorableInt(widget.counter);
+
+  @override
+  String? get restorationId => 'multiplication_page';
+
+  @override
+  void restoreState(RestorationBucket? oldBucket, bool initialRestore) {
+    registerForRestoration(_counter, 'multi_counter');
+  }
+
+  @override
+  void dispose() {
+    _counter.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    return const CupertinoPageScaffold(
-      backgroundColor: CupertinoColors.activeOrange,
-      child: SheetBody(title: 'Next Page')
+    return CupertinoPageScaffold(
+      child: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text('Current Count'),
+            Text(_counter.value.toString()),
+            CupertinoButton(
+              onPressed: () {setState(() => _counter.value = _counter.value * 2);},
+              child: const Text('Double it'),
+            ),
+            CupertinoButton(
+              onPressed: () => Navigator.pop(context, _counter.value),
+              child: const Text('Pass it on to the last sheet'),
+            ),
+          ]
+        ),
+      )
     );
   }
 }
